Alter Circle.yml to successfully build
CircleCI uses the non-recommended multi-directory GOPATH
(https://github.com/golang/go/wiki/GOPATH#use-a-single-gopath). We only
have write access to the .go_workspace directory anyway, so the GOPATH
for this build has been changed to just "$HOME/.go_workspace" to make
subsequent commands work. When CircleCI clones the repo, it is cloned
into the home directory, so we need to copy it into the GOPATH, which is
accomplished using rsync thanks to the recommendation of this Thoughtbot
post: https://robots.thoughtbot.com/configure-circleci-for-go.

Unfortunately this doesn't entirely work because the ".git" directory is
not included, so `go get` isn't able to ascertain which version control
system our project is using. Adding '--include=".git"' to `rsync` fixes
this and is effectively a local clone.

The repository will already be checked out to the appropriate branch
after `rsync`, however CircleCI doesn't setup remote tracking branches,
which `go get` needs to work.
Adding `git branch --set-upstream-to=origin/$CIRCLE_BRANCH` fixes this.